pyproject.toml
[project]
name = "snobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "duckdb>=1.3.2",
    "faiss-cpu>=1.11.0.post1",
    "opaiui>=0.13.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.48.0",
]
